Handled sending payload through pension adviser journey as well
